fix(steam): encode UpdateAuthSessionWithSteamGuardCode steamid as fixed64
The request proto declares steamid as fixed64 (wire type 1, 8-byte LE,
verified against the upstream compiled descriptors). Encoding it as
varint makes the server treat it as an unknown field and drop it
(steamid=0), answering EResult 8 InvalidParam — surfaced by this
morning's X-eresult check as
'UpdateAuthSessionWithSteamGuardCode failed (eresult=8 InvalidParam)'.

- encodeUpdateGuard: steamid via pbWriter.f64
- golden wire-type assertion TestEncodeUpdateGuardWireTypes
  (red->green, archived in evidence)
- all four auth Request messages re-audited field-type by field-type;
  PollAuthSessionStatus.token_to_revoke is the only other fixed64 and
  is an optional field we do not send
- fix(uu): commit Crypt.Key accessor — uu-probe (committed earlier)
  references it but the method existed only as an uncommitted
  working-tree edit, leaving the committed tree uncompilable
- api-notes §1.1: fixed64 annotation + per-request field-type table

+++ backend/internal/platform/steam/login.go
@@ -256,7 +256,7 @@ func encodePoll(clientID uint64, requestID []byte) []byte {
 func encodeUpdateGuard(clientID, steamID uint64, code string, codeType int32) []byte {
 	var w pbWriter
 	w.u64(1, clientID)
-	w.u64(2, steamID)
+	w.f64(2, steamID) // steamid is declared fixed64 in the request proto — varint makes the server drop it (EResult 8 InvalidParam)
 	w.str(3, code)
 	w.u64(4, uint64(codeType))
 	return w.buf
